+++ packages/core/src/odata-common/request-builder/count-request-builder.ts
@@ -6,7 +6,7 @@ import {
   DestinationNameAndJwt
 } from '../../connectivity/scp-cf';
 import { ODataCountRequestConfig } from '../request/odata-count-request-config';
-import { HttpRequestAndResponse } from '../../http-client';
+import { HttpResponse } from '../../http-client';
 import type { GetAllRequestBuilder } from './get-all-request-builder-base';
 
 /**
@@ -39,7 +39,7 @@ export class CountRequestBuilder<
     options?: DestinationOptions
   ): Promise<number> {
     return this.executeRaw(destination, options)
-      .then(({ response }) => {
+      .then((response) => {
         if (typeof response.data !== 'number') {
           throw new Error('Count request did not return a bare number.');
         }
@@ -48,17 +48,17 @@ export class CountRequestBuilder<
   }
 
   /**
-   * Execute request and return the request and the raw response.
+   * Execute request and return an [[HttpResponse]].
    *
    * @param destination - Destination to execute the request against
    * @param options - Options to employ when fetching destinations
-   * @returns A promise resolving to an [[HttpRequestAndResponse]].
+   * @returns A promise resolving to an [[HttpResponse]].
    */
   async executeRaw(
     destination: Destination | DestinationNameAndJwt,
     options?: DestinationOptions
-  ): Promise<HttpRequestAndResponse>{
+  ): Promise<HttpResponse>{
     return this.build(destination, options)
-      .then(request => request.executeRaw());
+      .then(request => request.execute());
   }
 }