+++ packages/core/src/odata-v4/request-builder/action-import-request-builder.spec.ts
@@ -92,9 +92,8 @@ describe('action import request builder', () => {
       const actual = await testActionImportNoParameterNoReturnType({}).executeRaw(
         destination
       );
-      expect(actual.response.data).toEqual({});
-      expect(actual.request.method).toBe('post');
-      expect(actual.request.baseURL).toBe(destination.url);
+      expect(actual.data).toEqual({});
+      expect(actual.request.method).toBe('POST');
     });
   });
 });